Generics {modeltools}	R Documentation

Generic Utility Functions

Description

A collection of standard generic functions for which other packages provide methods.

Usage

ICL(object, ...)
KLdiv(object, ...)
Lapply(object, FUN, ...)
clusters(object, newdata, ...)
getModel(object, ...)
parameters(object, ...)
posterior(object, newdata, ...)
prior(object, ...)
refit(object, newdata, ...)
relabel(object, by, ...)
ParseFormula(formula, data = list())

Arguments

object
S4 classed object.

formula
A model formula.

data
An optional data frame.

FUN
The function to be applied.

newdata
Optional new data.

by
Typically a character string specifying how to relabel the object.

...
Some methods for these generic function may take additional, optional arguments.

Details

ICL:

Integrated Completed Likelihood criterion for model selection.

KLdiv:

Kullback-Leibler divergence.

Lapply:

S4 generic for lapply

clusters:

Get cluster membership information from a model or compute it for new data.

getModel:

Get single model from a collection of models.

parameters:

Get parameters of a model (similar to but more general than coefficients).

posterior:

Get posterior probabilities from a model or compute posteriors for new data.

prior:

Get prior probabilities from a model.

refit:

Refit a model (usually to obtain additional information that was not computed or stored during the initial fitting process).

relabel:

Relabel a model (usually to obtain a new permutation of labels in mixture models or cluster objects).

Author(s)

Friedrich Leisch

[Package modeltools version 0.2-23 Index]