MEapply {modeltools}	R Documentation

Apply functions to Data in Object of Class "ModelEnv"

Description

Apply a single function or a collection of functions to the data objects stored in a model environment.

Usage

## S4 method for signature 'ModelEnv'
MEapply(object, FUN, clone = TRUE, ...)

Arguments

object
Object of class "ModelEnv".

FUN
Function or list of functions.

clone
If TRUE, return a clone of the original object, if FALSE, modify the object itself.

...
Passed on to FUN.

Examples

data("iris")
me <- ModelEnvFormula(Species+Petal.Width~.-1, data=iris,
                      subset=sample(1:150, 10))

me1 <- MEapply(me, FUN=list(designMatrix=scale,
                            response=function(x) sapply(x, as.numeric)))

me@get("designMatrix")
me1@get("designMatrix")

[Package modeltools version 0.2-23 Index]