Predict {modeltools}	R Documentation

Model Predictions

Description

A function for predictions from the results of various model fitting functions.

Usage

Predict(object, ...)

Arguments

object
a model object for which prediction is desired.

...
additional arguments affecting the predictions produced.

Details

A somewhat improved version of predict for models fitted with objects of class StatModel-class.

Value

Should return a vector of the same type as the response variable specified for fitting object.

Examples

df <- data.frame(x = runif(10), y = rnorm(10))
mf <- dpp(linearModel, y ~ x, data = df)
Predict(fit(linearModel, mf))

[Package modeltools version 0.2-23 Index]