info {modeltools}	R Documentation

Get Information on Fitted Objects

Description

Returns descriptive information about fitted objects.

Usage

info(object, which, ...)
## S4 method for signature 'ANY,missing'
info(object, which, ...)
infoCheck(object, which, ...)

Arguments

object
fitted object.

which
which information to get. Use which="help" to list available information.

...
passed to methods.

Details

Function info can be used to access slots of fitted objects in a portable way.

Function infoCheck returns a logical value that is TRUE if the requested information can be computed from the object.

Author(s)

Friedrich Leisch

[Package modeltools version 0.2-23 Index]